requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
statsmodels>=0.13.0
fredapi>=0.5.0
world-bank-data>=0.1.0
dash>=2.0.0
plotly>=5.0.0
tensorflow>=2.7.0
pmdarima>=2.0.0
pytest>=6.0.0
black>=21.5b2
flake8>=3.9.0
jupyter>=1.0.0
python-dotenv>=0.15.0 